Dodana metryka miejska i jej analiza
