+++ main.py
@@ -1,31 +1,21 @@
-from sklearn.neighbors import KNeighborsClassifier
-
-from clean_and_split_data import process_data
+from helpers import process_data, knn_analysis
 import pandas as pd
 import numpy as np
-from knn import KNN
-
-
-
 
 if __name__ == "__main__":
-    X_train, y_train, header = process_data("data.csv", "cleaned_data.csv", "train.csv", "test.csv")
+    X_train, X_test, header = process_data("data.csv", "cleaned_data.csv", "train.csv", "test.csv")
     X_train = np.array(X_train, dtype=float)
-    y_train = np.array(y_train, dtype=int)
-    df = pd.DataFrame(X_train, columns=header)
+    y_train = X_train[:, -1]
+    X_train = X_train[:, :-1]
+
+    X_test = np.array(X_test, dtype=int)
+    y_test = X_test[:, -1]
+    X_test = X_test[:, :-1]
+
+    df = pd.DataFrame(np.column_stack((X_train, y_train)), columns=header )
     print(df)
 
-for k in range(1,22):
-    knn = KNN(n_neighbours=k)
-    knn.fit(X_train[:, :-1], X_train[:, -1])
-    predictions = knn.predict(y_train[:, :-1])
-    print("________________________________")
-    accuracy = np.mean(predictions == y_train[:, -1])
-    print(f"Dokładność customowego: {accuracy:.4f}")
 
-    built_in_knn = KNeighborsClassifier(n_neighbors=k)
-    built_in_knn.fit(X_train[:, :-1], X_train[:, -1])
-    predictions = built_in_knn.predict(y_train[:, :-1])
+    knn_analysis(X_train, X_test, y_train, y_test, "manhattan")
+    knn_analysis(X_train, X_test, y_train, y_test, "euclidean")
 
-    accuracy = np.mean(predictions == y_train[:, -1])
-    print(f"Dokładność wbudowanego: {accuracy:.4f}")
